[Refactor] KeyChain -> KeyChainManager 변경
